requirements.txt
dash
web3==6.20.3
chart-builder
python-dotenv
flask
ipykernel
networkx
nbformat
apscheduler